+++ src/tcode_api/schemas/descriptions/labware/lid/latest.py
@@ -1,1 +1,1 @@
-from .v3 import LidDescription, LidDescriptor
+from .v4 import LidDescription, LidDescriptor