+++ src/tcode_api/schemas/descriptions/labware/well_plate/latest.py
@@ -1,1 +1,1 @@
-from .v3 import WellPlateDescription, WellPlateDescriptor
+from .v4 import WellPlateDescription, WellPlateDescriptor